feat(routes): add form definitions for various property and user routes
- Enhanced the properties routes with form definitions for city, index, create, edit, update, and destroy actions.
- Introduced form definitions for user registration, password management, two-factor authentication, and email verification routes.
- Each route now includes methods for handling GET, POST, and HEAD requests with appropriate action URLs and query options.

+++ app/Http/Controllers/MediaController.php
@@ -192,7 +192,7 @@ class MediaController extends Controller
 
         foreach ($request->media_ids as $mediaId) {
             $media = Media::find($mediaId);
-            if ($media && $media->user_id === auth()->id() && !$property->media()->where('media_id', $mediaId)->exists()) {
+            if ($media && (int) $media->user_id === (int) auth()->id() && !$property->media()->where('media_id', $mediaId)->exists()) {
                 $property->media()->attach($mediaId, ['order' => $property->media()->count() + 1]);
             }
         }
